backend/requirements.txt
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-mistralai
langchain-mcp-adapters
fastmcp
httpx 
bs4
tzdata
pyyaml
langgraph-checkpoint-postgres 
fastapi
langgraph
langgraph-cli
langgraph-cli[inmem]
pycryptodome
langgraph-checkpoint-firestore
google-cloud-firestore